fix: 14 medium-priority Ollama API compatibility gaps
M-1:  /api/ps adds size_vram, expires_at, model fields
M-2:  /api/create streams NDJSON status updates (stream=true default)
M-3:  Already safe - MessageContent::Text only in responses
M-4:  Already handled by backend
M-5:  /api/tags adds 'model' field to NativeModelInfo
M-6:  /api/show adds 'parent_model' extracted from Modelfile FROM
M-7:  /api/show 'parameters' returns key-value text, not JSON
M-8:  Blob handler errors return JSON instead of plain text
M-9:  Timestamps use Ollama format with nanoseconds and Z suffix
M-10: /api/embeddings response includes 'model' field
M-11: /api/embed response includes 'prompt_eval_count' field
M-12: Already compatible - returns {"version": "0.1.0"}
M-13: 'format' field accepts both string and JSON Schema object
M-14: 'seed' field widened from u32 to i64 (accepts -1)

All 435 tests pass.

+++ src/backend/types.rs
@@ -160,7 +160,7 @@ pub struct ChatRequest {
     #[serde(default)]
     pub presence_penalty: Option<f32>,
     #[serde(default)]
-    pub seed: Option<u32>,
+    pub seed: Option<i64>,
     #[serde(default)]
     pub num_ctx: Option<u32>,
     #[serde(default)]
@@ -228,7 +228,7 @@ pub struct CompletionRequest {
     #[serde(default)]
     pub presence_penalty: Option<f32>,
     #[serde(default)]
-    pub seed: Option<u32>,
+    pub seed: Option<i64>,
     #[serde(default)]
     pub num_ctx: Option<u32>,
     #[serde(default)]
